requirements.txt
Flask>=2.3.0
Flask-Session>=0.4.0
Flask-SQLAlchemy>=3.0.0
psycopg2-binary>=2.9.0
gunicorn
